test: messages/twitch-types/connection tests (spec => test to match Deno's pattern)
