+++ lib/message/twitch-types/connection/reconnect.test.ts
@@ -1,0 +1,12 @@
+import { parseTwitchMessage } from "../../parser/twitch-message.ts";
+import { ReconnectMessage } from "./reconnect.ts";
+import { assertInstanceOf } from "https://deno.land/x/explicitly/mod.ts";
+
+
+Deno.test("ReconnectMessage should be able to parse a real RECONNECT message", () => {
+  const msg = parseTwitchMessage(
+    ":tmi.twitch.tv RECONNECT",
+  ) as ReconnectMessage;
+
+  assertInstanceOf(msg, ReconnectMessage);
+});
